Change how PHP 7+ defines PMP_NOTIFICATIONS_SECRET, filtering the salt provided to crypt().
The reason that we must filter the salt value in PHP 7 is to make sure it's limited to the character set `./0-9A-Za-z`, since having any other character in the salt will cause crypt() to fail with the following message, as seen in Travis CI tests:

> Deprecated: crypt(): Supplied salt is not valid for DES. Possible bug in provided salt format. in /tmp/wordpress/src/wp-content/plugins/pmp-wordpress-plugin/inc/notifications.php on line 3

PHP 5 environments do not have this limitation.

This implements the "hybrid" approach described at https://github.com/npr/pmp-wordpress-plugin/issues/133#issuecomment-384469661, for https://github.com/npr/pmp-wordpress-plugin/issues/133.