Reorder .travis.yml test matrix to improve pull-request turnaround time
PHP 5 tests have been fairly uniformly passing. At 5 minutes per test run for each of the 9 PHP 5 tests run by Travis, it takes about 45 minutes from when I submit a pull request aginst GitHub to when the first of the PHP 7 tests start to run. The PHP 7 tests have been having test failures recently. This means that there's a 45-minute gap between when a pull request triggers a test and when the test starts to deliver useful data.

Putting the PHP 7 test first puts the more-likely-to-fail tests earlier, decreasing the amount of time that a pull request submitter must wait for feedback.