Allow sites to unset their PMP Secret without resorting to database edits.
This implements the "alternate solution" described in
https://github.com/npr/pmp-wordpress-plugin/issues/130,
whereby the PMP client secret's value is never passed to the browser,
but the browser can unset the PMP client secret's value by passing
an empty value with the form while also passing a checked box stating
intent to change the value.

The conditional in this is gnarly, and therefore well-commented.